==> Fig4_Amin_mapping/02_NOMS_to_HDBCA_mapping.ipynb <==
# mapping pasca to braun

import warnings

warnings.filterwarnings("ignore")

import os

import pandas as pd

import scvi
from scvi.model.utils import mde

import scanpy as sc

DATA_DIR = "/storage/data/data/organoid_atlas/"

ANNOT_LABEL = "subregion_class"

adata_q = sc.read(os.path.join(DATA_DIR, "230605_pasca_all_v1_common_hv2k.h5ad"))

adata_ref = sc.read(os.path.join("/storage/data/data/organoid_atlas/jonas_query_ref_latent/braun_2022_fetal_brain_v2_common_hv2k.h5ad"))

# try with orig.ident vs dataset
# include only primary/secondary

# subset adata_q to organoid
adata_q = adata_q[~adata_q.obs.dataset.isin(["fetal_striatum_21pcw"])]

adata_q.varm = dict()
adata_q.obs["batch"] = adata_q.obs["dataset"].astype(str).copy()

adata_ref.varm = dict()
adata_ref.obs["batch"] = adata_ref.obs["Donor"].astype(str).copy()

lvae = scvi.model.SCANVI.load(
        os.path.join(DATA_DIR, "query_scarches_nolabs"),
        adata=adata_ref,
    )

adata_q.obs[ANNOT_LABEL] = "Unknown"

scvi.model.SCANVI.prepare_query_anndata(adata_q, lvae)

vae_q = scvi.model.SCANVI.load_query_data(adata_q, lvae)

vae_q.train(
    max_epochs=500,
    plan_kwargs=dict(weight_decay=0.0),
)

adata_q.obs["ref"] = "query"
adata_ref.obs["ref"] = "reference"

adata_full = adata_q.concatenate(adata_ref)

adata_full.obs["batch"] = (
    adata_q.obs["batch"].tolist() + adata_ref.obs["batch"].tolist()
)

adata_full.obsm["X_scARCHES"] = vae_q.get_latent_representation(adata_full)
adata_full.obsm["X_umap"] = mde(adata_full.obsm["X_scARCHES"])

p = sc.pl.umap(
    adata_full,
    color=["batch", "ref"],
    show=False,
)

p[0].figure.savefig(os.path.join(DATA_DIR,f"scanvi_full_mde.png"))

full_latent = pd.DataFrame(adata_full.obsm["X_scARCHES"])
full_latent.index = adata_full.obs.index

full_latent.to_csv(os.path.join(DATA_DIR, f"pasca_map_braun_latent.csv"))

vae_q.save(
    os.path.join(DATA_DIR, "pasca_map_braun_model.pt"),
    overwrite=True,
    save_anndata=False,
)
