==> Fig4_Amin_mapping/03_NOMS_to_HNOCA_mapping.ipynb <==
import pandas as pd
import scanpy as sc
import numpy as np

from scarches.models.scpoli import scPoli

scpoli_model_path = "/storage/data/data/organoid_atlas/230510_scpoli_hierarchical123"
ref_adata_path = "/storage/data/data/organoid_atlas/230510_09_organoids_cleaned_hvg.h5ad"
query_adata_path = "/storage/data/data/organoid_atlas/230605_pasca_all_01.h5ad"

ref_adata = sc.read_h5ad(ref_adata_path)
ref_adata.X = ref_adata.layers["counts_lengthnorm"].A.copy()

scpoli_model = scPoli.load(scpoli_model_path, ref_adata)

query_adata = sc.read_h5ad(query_adata_path)

query_adata.obs["batch"] = query_adata.obs["dataset"]

query_adata.obs["snapseed_pca_rss_level_1"] = "unknown"
query_adata.obs["snapseed_pca_rss_level_12"] = "unknown"
query_adata.obs["snapseed_pca_rss_level_123"] = "unknown"

query_adata.X = query_adata.layers["counts"].A.copy()

query_adata = query_adata[:, ref_adata.var.index]

scpoli_query = scPoli.load_query_data(
    adata=query_adata,
    reference_model=scpoli_model,
    labeled_indices=[]
)

scpoli_query.train(
    n_epochs=500,
    pretraining_epochs=500,
    eta=10,
    unlabeled_prototype_training=False
)

query_data_latent = scpoli_query.get_latent(
    query_adata,
    mean=True
)

ref_data_latent = scpoli_model.get_latent(
    ref_adata,
    mean=True
)

joint_latent = np.concatenate((ref_data_latent, query_data_latent), axis=0)

# create new adata
joint_adata = sc.AnnData(
    X=joint_latent,
    obs=pd.concat([ref_adata.obs, query_adata.obs], axis=0)
)

joint_adata.obs["query"] = False

indices = range(-len(query_data_latent), 0)
joint_adata.obs.loc[joint_adata.obs.iloc[indices].index, "query"] = True

sc.pp.neighbors(joint_adata)
sc.tl.umap(joint_adata)

joint_adata.obs["query"] = joint_adata.obs["query"].astype(str)

joint_adata.obs["conditions_combined"] = joint_adata.obs["conditions_combined"].astype(str)

joint_adata.write_h5ad("/storage/data/data/organoid_atlas/scpoli_scarches_pasca.h5ad")

sc.pl.umap(
    joint_adata,
    color='snapseed_pca_rss_level_1',
    show=False,
    frameon=False,
    save="_scanpseed.png"
)

sc.pl.umap(
    joint_adata,
    color='snapseed_pca_rss_level_12',
    show=False,
    frameon=False,
    save="_scanpseed_level12.png"
)

sc.pl.umap(
    joint_adata,
    color='snapseed_pca_rss_level_123',
    show=False,
    frameon=False,
    save="_scanpseed_level123.png"
)

sc.pl.umap(
    joint_adata,
    color='class',
    show=False,
    frameon=False,
    save="_pasca_class.png"
)

sc.pl.umap(
    joint_adata,
    color='class2',
    show=False,
    frameon=False,
    save="_pasca_class2.png"
)

sc.pl.umap(
    joint_adata,
    color='region',
    show=False,
    frameon=False,
    save="_pasca_region.png"
)
